finsight/backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
sqlalchemy==2.0.23
flask-cors==4.0.0
flask-login==0.6.3
bcrypt==4.0.1
cryptography==41.0.7
python-dotenv==1.0.0
reportlab==4.0.4
fpdf2==2.7.6
jinja2==3.1.2
pdfkit==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dateutil==2.8.2
google-generativeai==0.3.2
asyncio==3.4.3
aiohttp==3.9.1
